backend/models/train_random_forest.py
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, precision_score, recall_score, roc_auc_score, f1_score
import joblib

# Load processed data
df = pd.read_csv("data/processed/credit_data_processed.csv")

X = df.drop(columns=["default_flag"])
y = df["default_flag"]

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42, stratify=y
)

model = RandomForestClassifier(
    n_estimators=200,
    max_depth=10,
    class_weight="balanced",
    random_state=42
)

model.fit(X_train, y_train)

y_pred = model.predict(X_test)
y_prob = model.predict_proba(X_test)[:, 1]

print("\n🌲 RANDOM FOREST RESULTS")
print("------------------------")
print("Accuracy :", accuracy_score(y_test, y_pred))
print("Precision:", precision_score(y_test, y_pred))
print("Recall   :", recall_score(y_test, y_pred))
print("ROC-AUC  :", roc_auc_score(y_test, y_prob))
print("F1 Score :", f1_score(y_test, y_pred))

joblib.dump(model, "models/random_forest_model.pkl")

print("\n✅ Random Forest model trained and saved!")

